refactor: sort features by category-folders
