+++ src/app/catalog/_components/ObjectDetail/ObjectHero.tsx
@@ -5,8 +5,8 @@ import Section from '@/layout/Section'
 import MapObjectsButton from '@/ui/buttons/MapObjectsButton'
 import PropItem from '@/app/catalog/_components/ObjectDetail/PropItem'
 import Button from '@/ui/buttons/Button'
-import { formatDate } from '@/features/date'
-import { formatFullPrice, formatPriceForArea } from '@/features/price'
+import { formatDate } from '@/features/utility/date'
+import { formatFullPrice, formatPriceForArea } from '@/features/utility/price'
 import { OBJECT_PROPS_ID } from '@/app/catalog/_components/ObjectDetail/ObjectProperties/ObjectProperties'
 
 import styles from './ObjectHero.module.css'
