+++ src/app/catalog/complexes/[complex]/_components/ComplexHero.tsx
@@ -4,7 +4,7 @@ import { DetailComplex } from '@/types/Complex'
 import Img from '@/ui/Img'
 import Image from 'next/image'
 import DeveloperLogo from '@/assets/static/donstroy.svg'
-import { formatPriceShortBy } from '@/features/price'
+import { formatPriceShortBy } from '@/features/utility/price'
 import Button from '@/ui/buttons/Button'
 import NoPhoto from '@/assets/static/no-photo.jpg'
 import { DecorativeBlock } from '@/layout/DecorativeSection'
