+++ src/app/catalog/complexes/page.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { getComplexes } from '@/globals/api'
 import CatalogContent from '@/layout/catalog/CatalogContent'
 import CategoryLayout from '@/layout/catalog/CategoryLayout'
-import { GetObjectsFn } from '@/features/useCategoryObject'
+import { GetObjectsFn } from '@/features/catalog/useCategoryObject'
 import { ComplexCard as ComplexCardType } from '@/types/Complex'
 
 async function Page() {
