+++ src/components/QuickFilter/QuickFilter.tsx
@@ -3,7 +3,7 @@ import DetailFilterButton from '@/components/QuickFilter/DetailFilterButton'
 import Container from '@/layout/Container'
 import MapObjectsButton from '@/ui/buttons/MapObjectsButton'
 import CatalogViewButton from '@/components/QuickFilter/CatalogViewButton'
-import { flatPlural } from '@/features/pluralRules'
+import { flatPlural } from '@/features/utility/pluralRules'
 import { QuickFilters } from '@/types/FiltersType'
 import { FilterItems } from '@/features/FilterItems'
 import SortButton from '@/components/QuickFilter/SortButton'
