+++ src/features/utility/pluralRules.ts
