+++ src/layout/catalog/CategoryLayout.tsx
@@ -3,7 +3,7 @@ import Breadcrumbs from '@/components/Breadcrumbs'
 import { AnyCategory } from '@/types/Category'
 import PageTitle from '@/components/PageTitle/PageTitle'
 import Container from '@/layout/Container'
-import { suggestionPlural } from '@/features/pluralRules'
+import { suggestionPlural } from '@/features/utility/pluralRules'
 import QuickFilter from '@/components/QuickFilter/QuickFilter'
 import { CategoryProvider } from '@/layout/catalog/CategoryContext'
 import SubCategoryLink from '@/components/SubCategoryLink'
