+++ src/ui/cards/ObjectCard/ObjectListCard.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { ObjectCard } from '@/types/ObjectCard'
 import Link from 'next/link'
 import { ActionButton } from '@/ui/cards/ObjectCard/ActionButton'
-import { CategoryForGeneratingLink, generateObjectLink } from '@/features/link'
+import { CategoryForGeneratingLink, generateObjectLink } from '@/features/catalog/link'
 import { LayoutObject } from '@/types/Complex'
 import Gallery from '@/ui/cards/ObjectCard/Gallery'
 import Tags from '@/components/Tags'
