+++ src/ui/cards/ObjectCard/ObjectTileCard.tsx
@@ -2,7 +2,7 @@ import React, { PropsWithChildren } from 'react'
 import Link from 'next/link'
 import { ObjectCard as ObjectCardType } from '@/types/ObjectCard'
 import { ActionButton } from '@/ui/cards/ObjectCard/ActionButton'
-import { CategoryForGeneratingLink, generateObjectLink } from '@/features/link'
+import { CategoryForGeneratingLink, generateObjectLink } from '@/features/catalog/link'
 import { LayoutObject } from '@/types/Complex'
 import Gallery from '@/ui/cards/ObjectCard/Gallery'
 import Tags from '@/components/Tags'
