+++ src/ui/popups/FilterPopup/FilterPopup.tsx
@@ -5,7 +5,7 @@ import ClosePopupButton from '@/ui/buttons/ClosePopupButton'
 import Button from '@/ui/buttons/Button'
 import { getFilters } from '@/globals/api/methods/getFilters'
 import { FilterBlock } from '@/types/FiltersType'
-import { IsDesktop, IsMobile } from '@/features/adaptive'
+import { IsDesktop, IsMobile } from '@/features/utility/adaptive'
 import { FilterItems } from '@/features/FilterItems'
 import ResetFiltersButton from '@/components/QuickFilter/ResetFiltersButton'
 import { FiltersTags } from '@/components/QuickFilter/FilterTags'
